% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cards.R
\name{add_wild_cards}
\alias{add_wild_cards}
\title{Add wild cards to a hand, replacing jokers}
\usage{
add_wild_cards(hand, cards)
}
\arguments{
\item{hand}{A hand}

\item{cards}{Cards to add}
}
\value{
A new hand with the cards added
}
\description{
Add wild cards to a hand, replacing jokers
}
