% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{assignToBoard}
\alias{assignToBoard}
\title{Function: 	assignToBoard}
\usage{
assignToBoard(y)
}
\description{
Author: Ben Greenspan		
Created: May/June 2010
Description:	deal 3 community cards 
Inputs:		@param y : cards dealt					#vector[2*nPlayers+3] in {1, 2, ..., 52}
Outputs:		@return board :	the board cards			#vector[5] in {1, 2, ..., 52}
}
