% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{assignToPlayers}
\alias{assignToPlayers}
\title{Function: 	assignToPlayers}
\usage{
assignToPlayers(nPlayers, position, y)
}
\description{
Author: Ben Greenspan		
Created: May/June 2010
Description:	standard normal deal: begin the deal starting at the left of the dealer.
Inputs:		@param nPlayers : number of hands to deal 
											#integer in {1, 2, ... , 8}
			position : dealer position 	#integer in {1, 2, ..., nPlayers}
			@param y : cards dealt					#vector[2*nPlayers+5] in {1, 2, ..., 52}
Locals:		i : for name in vector
											#numeric in vector
			k : for name in vector
											#numeric in vector
Outputs:		@return players :	the hole cards in absolute position
											#matrix[nPlayers, 4]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: suit of card 1 in {1, 2, 3, 4}
											#col3: rank of card 2
											#col4: suit of card 2
}
