% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hands.R
\name{categorize_hand}
\alias{categorize_hand}
\title{Find the category of a hand.
Will work with hands up to 9 cards.}
\usage{
categorize_hand(hand)
}
\arguments{
\item{hand}{A hand}
}
\value{
The hand with added category and rank_vector fields
}
\description{
Find the category of a hand.
Will work with hands up to 9 cards.
}
