% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{CGIplayers}
\alias{CGIplayers}
\title{CGIplayers}
\usage{
CGIplayers(time, alias, position, cards, score)
}
\arguments{
\item{time}{\tabular{ll}{the current round as    integer in \{1, 2, 3, 4\} \tab \cr \tab
1 = pre-flop \cr \tab
2 = flop \cr \tab
3 = turn \cr \tab
4 = river}}

\item{alias}{names of players    as                vector[nPlayers]}

\item{position}{dealer position as             integer in \{2, ..., nPlayers\}}

\item{cards}{\tabular{ll}{the 7 card hand as
matrix[nPlayers, 14] \tab \cr \tab
col1: rank of card 1 in \{2, ... , 14\} \cr \tab
col2: suit of card 1 in \{1, 2, 3, 4\} \cr \tab
col3: rank of card 2 \cr \tab
col4: suit of card 2 \cr \tab
. \cr \tab
. \cr \tab
. \cr \tab
col13: rank of card 7 \cr \tab
col14: suit of card 7}}

\item{score}{\tabular{ll}{the score of the hand in absolute terms as
vector[nPlayers] \tab \cr \tab
9 = Straight Flush \cr \tab
8 = Four of a Kind \cr \tab
7 = Full House \cr \tab
6 = Flush \cr \tab
5 = Straight \cr \tab
4 = Three of a Kind \cr \tab
3 = Two Pair \cr \tab
2 = One Pair \cr \tab
1 = High Card}}
}
\value{
In lieu of a return value, CGIplayers calls the plot() function.
}
\description{
A primitive method (i.e., does not support classes) for graphics using the plot() function. Built-in support for 2-9 players.  This function was written on a Mac and may not be PC-compatible (yet). You must have already called CGIplayers(time=1, ...) before calling CGIplayers(time=2, ...), you must have already called CGIplayers(time=1, ...) and CGIplayers(time=2, ...) before calling CGIplayers(time=3, ...), etc.
}
\details{
Locals:\cr
nPlayers : number of hands as			integer in \{2, ... , 9\} \cr
width : width of the plot	as			integer in {20, 30, ... , 10*nPlayers} \cr
height : height of the plot as		integer in {10} \cr
X1 : x-coordinate for Players	as		real number in [0, width] \cr
Y1 : y-coordinate for Players	as		real number in [0, height] \cr
X2 : x-coordinate for Cards as		real number in [0, width] \cr
Y2 : y-coordinate for Cards as		real number in [0, height] \cr
X3 : x-coordinate for Community as	real number in [0, width] \cr
Y3 : y-coordinate for Community as	real number in [0, height] \cr
X4 : x-coordinate for outline as		real number in [0, width] \cr
Y4 : y-coordinate for outline as		real number in [0, height] \cr
}
\examples{
CGIplayers(1,c("Player 1", "Player 2","Player 3", "Player 4", "Player 5", "Player 6", "Player 7", "Player 8", "Player 9"),9,matrix(c(2,8,12,14,10,6,14,8,4,2,3,2,4,1,4,3,1,1,13,4,4,5,3,9,8,12,7,3,4,3,2,2,4,2,1,1,3,3,3,3,3,3,3,3,3,1,1,1,1,1,1,1,1,1,10,10,10,10,10,10,10,10,10,4,4,4,4,4,4,4,4,4,12,12,12,12,12,12,12,12,12,4,4,4,4,4,4,4,4,4,11,11,11,11,11,11,11,11,11,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,3,3,3,3,3,3,3,3,3),nrow=9,ncol=14,byrow=FALSE),c(2,1,2,1,3,1,1,2,1))
CGIplayers(2,c("Player 1", "Player 2","Player 3", "Player 4", "Player 5", "Player 6", "Player 7", "Player 8", "Player 9"),9,matrix(c(2,8,12,14,10,6,14,8,4,2,3,2,4,1,4,3,1,1,13,4,4,5,3,9,8,12,7,3,4,3,2,2,4,2,1,1,3,3,3,3,3,3,3,3,3,1,1,1,1,1,1,1,1,1,10,10,10,10,10,10,10,10,10,4,4,4,4,4,4,4,4,4,12,12,12,12,12,12,12,12,12,4,4,4,4,4,4,4,4,4,11,11,11,11,11,11,11,11,11,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,3,3,3,3,3,3,3,3,3),nrow=9,ncol=14,byrow=FALSE),c(2,1,2,1,3,1,1,2,1))
CGIplayers(3,c("Player 1", "Player 2","Player 3", "Player 4", "Player 5", "Player 6", "Player 7", "Player 8", "Player 9"),9,matrix(c(2,8,12,14,10,6,14,8,4,2,3,2,4,1,4,3,1,1,13,4,4,5,3,9,8,12,7,3,4,3,2,2,4,2,1,1,3,3,3,3,3,3,3,3,3,1,1,1,1,1,1,1,1,1,10,10,10,10,10,10,10,10,10,4,4,4,4,4,4,4,4,4,12,12,12,12,12,12,12,12,12,4,4,4,4,4,4,4,4,4,11,11,11,11,11,11,11,11,11,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,3,3,3,3,3,3,3,3,3),nrow=9,ncol=14,byrow=FALSE),c(2,1,2,1,3,1,1,2,1))
CGIplayers(3,c("Player 1", "Player 2","Player 3", "Player 4", "Player 5", "Player 6", "Player 7", "Player 8", "Player 9"),9,matrix(c(2,8,12,14,10,6,14,8,4,2,3,2,4,1,4,3,1,1,13,4,4,5,3,9,8,12,7,3,4,3,2,2,4,2,1,1,3,3,3,3,3,3,3,3,3,1,1,1,1,1,1,1,1,1,10,10,10,10,10,10,10,10,10,4,4,4,4,4,4,4,4,4,12,12,12,12,12,12,12,12,12,4,4,4,4,4,4,4,4,4,11,11,11,11,11,11,11,11,11,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,3,3,3,3,3,3,3,3,3),nrow=9,ncol=14,byrow=FALSE),c(2,1,2,1,3,1,1,2,1))
CGIplayers(4,c("Player 1", "Player 2","Player 3", "Player 4", "Player 5", "Player 6", "Player 7", "Player 8", "Player 9"),9,matrix(c(2,8,12,14,10,6,14,8,4,2,3,2,4,1,4,3,1,1,13,4,4,5,3,9,8,12,7,3,4,3,2,2,4,2,1,1,3,3,3,3,3,3,3,3,3,1,1,1,1,1,1,1,1,1,10,10,10,10,10,10,10,10,10,4,4,4,4,4,4,4,4,4,12,12,12,12,12,12,12,12,12,4,4,4,4,4,4,4,4,4,11,11,11,11,11,11,11,11,11,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,3,3,3,3,3,3,3,3,3),nrow=9,ncol=14,byrow=FALSE),c(2,1,2,1,3,1,1,2,1))
}
