% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{deal}
\alias{deal}
\title{Function: 	deal}
\usage{
deal(nPlayers, position)
}
\description{
Author: Ben Greenspan		
Created: May/June 2010
Description:	generate Player+Community cards = 2x(nPlayers)+5
Inputs:		@param nPlayers : number of hands to deal 
											#integer in {1, 2, ... , 8}
			@param position : dealer position 	#integer in {1, 2, ..., nPlayers}
Locals:		nCards : number of cards dealt to players
											#integer in {4, 6, 8, ... 16}
Outputs:		@return y : cards dealt in hole			#vector[nCards] in {1, 2, ..., 52}
}
