% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_hand.R
\name{deal_hand}
\alias{deal_hand}
\title{Simulates dealing a 5-card poker hand.}
\usage{
deal_hand()
}
\value{
Generates a random five-card poker hands.
}
\description{
Simulates dealing a 5-card poker hand.
}
\examples{
poker_hand()
}
