% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotFlush}
\alias{dotFlush}
\title{Function: 	dotFlush}
\usage{
dotFlush(cards, score)
}
\description{
Author: Ben Greenspan		
Created: December 14, 2010
Updated: December 17, 2010
Description:	determines the player with the highest flush
Inputs:		@param cards : the 7 card hand
											#matrix[nPlayers, 14]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: suit of card 1 in {1, 2, 3, 4}
											#col3: rank of card 2
											#col4: suit of card 2
											#.
											#.
											#.
											#col13: rank of card 7
											#col14: suit of card 7
			@param score : the score of the hand in absolute terms
											#vector[nPlayers]
											# 9 = Straight Flush
											# 8 = Four of a Kind
											# 7 = Full House
											# 6 = Flush
											# 5 = Straight
											# 4 = Three of a Kind
											# 3 = Two Pair
											# 2 = One Pair
											# 1 = High Card
Locals:	flushRank : the rank of 5 high cards in flush
											#matrix[nPlayers,5]
											#col1: suit of card 1 in {2, ... , 14}
											#.
											#.
											#.
											#col5: suit of card n in {2, ... , 14}
}
