% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotFlushRanker}
\alias{dotFlushRanker}
\title{Function: 	dotFlushRanker}
\usage{
dotFlushRanker(cardsRow)
}
\description{
Author: Ben Greenspan		
Created: December 14, 2010
Updated: December 17, 2010
Description:	returns the rank of top 5 cards in the flush
Inputs:		@param cardsRow : one 7 card hand
											#vector[14]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: suit of card 1 in {1, 2, 3, 4}
											#col3: rank of card 2
											#col4: suit of card 2
											#.
											#.
											#.
											#col13: rank of card 7
											#col14: suit of card 7
Locals:		suits : the indices for suit	#vector
				sortedSuits : the sorted suits
											#vector
			suitValues : the unique values of equal suits
											#vector
											#col1: lowest unique suit
											#.
											#.
											#.
											#coln: highest unique suit
			suitLengths : the lengths of equal suits
											#vector
											#col1: length of lowest rank
											#.
											#.
											#.
											#coln: length of highest rank
			flushSuit : the suit of the flush
											#integer
			flushIndex : the indices of cards in flush
											#vector of integers 
											#col1: index in {1,2,...,7}
											#.
											#.
											#.
											#coln: index in {1,2,...,7}
Outputs:		@return flushRank : the rank of 5 high cards in flush
											#vector[5]
											#col1: suit of card 1 in {2, ... , 14}
											#.
											#.
											#.
											#col5: suit of card n in {2, ... , 14}
Outputs:		@return winner : absolute position of the winner
											#vector
}
