% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotFourOfAKind}
\alias{dotFourOfAKind}
\title{Function: 	dotFourOfAKind}
\usage{
dotFourOfAKind(nPlayers, cards, score)
}
\description{
Author: Ben Greenspan		
Created: December 15, 2010
Updated: December 17, 2010
Description:	determines the player with the four of a kind
Inputs:		@param nPlayers : number of hands		#integer in {1, 2, ... , 8}
			@param cards : the 7 card hand
											#matrix[nPlayers, 14]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: suit of card 1 in {1, 2, 3, 4}
											#col3: rank of card 2
											#col4: suit of card 2
											#.
											#.
											#.
											#col13: rank of card 7
											#col14: suit of card 7
			@param score : the score of the hand in absolute terms
											#vector[nPlayers]
											# 9 = Straight Flush
											# 8 = Four of a Kind
											# 7 = Full House
											# 6 = Flush
											# 5 = Straight
											# 4 = Three of a Kind
											# 3 = Two Pair
											# 2 = One Pair
											# 1 = High Card
Locals:		ranks : the indices for rank	#vector
			temp : a copy of cards with ranks only
											#matrix[nPlayers, 5]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: rank of card 2
											#.
											#.
											#.
											#col7: rank of card 7
			fourOfAKindRanks : a matrix for the rank of the quads and the kicker
											#matrix[nPlayers,2]
											#col1: rank of high set
											#col2: rank of high pair
Outputs:		@return winner : absolute position of the winner
											#vector
}
