% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotFullHouseRanker}
\alias{dotFullHouseRanker}
\title{Function: 	dotFullHouseRanker}
\usage{
dotFullHouseRanker(oneHand)
}
\description{
Author: Ben Greenspan		
Created: December 15, 2010
Description:	returns the rank of the top set and the top pair
Inputs:		@param oneHand : the ranks of one 7 card hand
											#vector[7]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: rank of card 2
											#.
											#.
											#.
											#col7: rank of card 7
Locals:		values : the unique values of equal ranks
											#vector
											#col1: top unique rank
											#.
											#.
											#.
											#coln: bottom unique rank
			lengths : the lengths of equal ranks
											#vector
											#col1: length of top rank
											#.
											#.
											#.
											#coln: length of bottom rank
Outputs:		@return fullHouseRank : the ranks of the high set and the high pair
											#vector
											#col1: the rank of the top set
											#col2: the rank of the top pair
}
