% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotHighcard}
\alias{dotHighcard}
\title{Function: 	dotHighcard}
\usage{
dotHighcard(cards)
}
\description{
Author: Ben Greenspan		
Created: December 14, 2010
Updated: December 17, 2010
Description:	determines the players with a high card hand
Inputs:		@param cards : the 7 card hand
											#matrix[nPlayers, 14]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: suit of card 1 in {1, 2, 3, 4}
											#col3: rank of card 2
											#col4: suit of card 2
											#.
											#.
											#.
											#col13: rank of card 7
											#col14: suit of card 7
Locals:		ranks : the indices for rank	#vector
			temp : a copy of cards with ranks only
											#matrix[nPlayers, 7]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: rank of card 2
											#.
											#.
											#.
											#col7: rank of card 7
			step : the step for comparison of highest card
											#integer
			Tie : the positions of the players with Tied highest cards
											#vector
Outputs:		@return winner : absolute position of the winner
											#vector
}
