% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotHighcardCompare}
\alias{dotHighcardCompare}
\title{Function: 	dotHighcardCompare}
\usage{
dotHighcardCompare(rankMatrix)
}
\description{
Author: Ben Greenspan		
Created: December 17, 2010
Description:	determines the players with the high card
Inputs:		@param rankMatrix : the ranks from the 7 card hand
											#matrix[nPlayers, 7]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: rank of card 2
											#.
											#.
											#.
											#col7: rank of card 7
Locals:		step : the step for comparison of highest card
											#integer
			Tie : the positions of the players with Tied highest cards
											#vector
Outputs:		@return winner : absolute position of the winner
											#vector
}
