% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotStraightFlushRanker}
\alias{dotStraightFlushRanker}
\title{Function: 	dotStraightFlushRanker
Author: Ben Greenspan		
Created: December 17, 2010
Description:	returns the top card in the straight flush
Inputs:		@param yTempRow : a sorted 7 card hand of numbers
											#vector[7]
											#col1: number of card 1 in {1, 2, ... , 52}
											#col2: number of card 2
											#.
											#.
											#.
											#col7: number of card 7
Locals:		
Outputs:		@return straightFlushRank : the top card in the straight flush
											#integer}
\usage{
dotStraightFlushRanker(yTempRow)
}
\description{
Function: 	dotStraightFlushRanker
Author: Ben Greenspan		
Created: December 17, 2010
Description:	returns the top card in the straight flush
Inputs:		@param yTempRow : a sorted 7 card hand of numbers
											#vector[7]
											#col1: number of card 1 in {1, 2, ... , 52}
											#col2: number of card 2
											#.
											#.
											#.
											#col7: number of card 7
Locals:		
Outputs:		@return straightFlushRank : the top card in the straight flush
											#integer
}
