% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotStraightRanker}
\alias{dotStraightRanker}
\title{Function: 	dotStraightRanker
Author: Ben Greenspan		
Created: December 14, 2010
Updated: December 17, 2010
Description:	returns the highest card in the straight
Inputs:		@param oneHand : a hand sorted from low to high with score of 5
											#vector[7]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: rank of card 2
											#.
											#.
											#.
											#col7: rank of card 7
Locals:		k : length of unique ranks 		#integer in {2, 3, ... , 7}
			counter : number of cards in sequence
Outputs:		@return straightRank : the top card in the straight
											#integer}
\usage{
dotStraightRanker(oneHand)
}
\description{
Function: 	dotStraightRanker
Author: Ben Greenspan		
Created: December 14, 2010
Updated: December 17, 2010
Description:	returns the highest card in the straight
Inputs:		@param oneHand : a hand sorted from low to high with score of 5
											#vector[7]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: rank of card 2
											#.
											#.
											#.
											#col7: rank of card 7
Locals:		k : length of unique ranks 		#integer in {2, 3, ... , 7}
			counter : number of cards in sequence
Outputs:		@return straightRank : the top card in the straight
											#integer
}
