% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotTestDealer}
\alias{dotTestDealer}
\title{Function: 	dotTestDealer}
\usage{
dotTestDealer(nPlayers, position, holeCards)
}
\description{
Author: Ben Greenspan		
Created: December 19, 2010
Description:	assume player 1 already has cards.  for remaining players, 
			generate Player+Community cards = 2x(nPlayers-1)+5
Inputs:		@param nPlayers : number of hands to deal 
											#integer in {1, 2, ... , 8}
			@param position : dealer position 	#integer in {1, 2, ..., nPlayers}
			@param holeCards : the hand of player 1
Locals:		nCards : number of cards dealt to players
											#integer in {4, 6, 8, ... 16}
Outputs:		@return y : cards dealt in hole			#vector[nCards] in {1, 2, ..., 52}
}
