% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotTransformToAbsolute}
\alias{dotTransformToAbsolute}
\title{dotTransformToAbsolute}
\usage{
dotTransformToAbsolute(nPlayers, position, k)
}
\arguments{
\item{nPlayers}{number of hands to deal as
integer in \{2, ... , 9\}}

\item{position}{dealer position as        integer in \{2, ..., nPlayers\}}

\item{k}{relative position of a player as
integer in \{1, 2, ... , nPlayers\}}
}
\value{
j : absolute position of a player as
											integer in \{1, 2, ... , nPlayers\}
}
\description{
Transform a relative position (i.e., seats behind the dealer) into an absolute position (i.e., seat at the table).
}
\examples{
dotTransformToAbsolute(9,9,0)
dotTransformToAbsolute(9,9,8)
dotTransformToAbsolute(9,1,8)
dotTransformToAbsolute(9,5,6)
}
