% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotTransformToRank}
\alias{dotTransformToRank}
\title{dotTransformToRank}
\usage{
dotTransformToRank(y)
}
\arguments{
\item{y}{number corresponding to card as
integer in \{1, 2, ... , 52\} \cr}
}
\value{
\tabular{ll}{rank: rank of card y as  
											integer in \{2, ... , 14\} \tab \cr \tab
											 2 = deuce \cr \tab
											 . \cr \tab
											 . \cr \tab
											 . \cr \tab
											 11 = jack \cr \tab
											 12 = queen \cr \tab
											 13 = king \cr \tab
											 14 = ace}
}
\description{
Determine the rank of a card.
}
\examples{
dotTransformToRank(1)
dotTransformToRank(13)
dotTransformToRank(14)
dotTransformToRank(26)
}
