% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotTransformToSuit}
\alias{dotTransformToSuit}
\title{dotTransformToSuit}
\usage{
dotTransformToSuit(y)
}
\arguments{
\item{y}{number corresponding to card as
integer in \{1, 2, ... , 52\}}
}
\value{
\tabular{ll}{suit: suit of card y as
											integer in \{1, 2, 3, 4\} \tab \cr \tab
											 1 = spade \cr \tab
											 2 = club \cr \tab
											 3 = heart \cr \tab
											 4 = diamond}
}
\description{
Determine the suit of a card.
}
\examples{
dotTransformToSuit(1)
dotTransformToSuit(13)
dotTransformToSuit(14)
dotTransformToSuit(26)
}
