% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalHand.R
\name{evalHand}
\alias{evalHand}
\title{Evaluate a poker hand}
\usage{
evalHand(x)
}
\arguments{
\item{x}{an integer matrix, where each column is a hand, or an integer vector.}
}
\value{
an integer, where higher = better
}
\description{
Returns an integer, where higher = a better hand
}
\details{
No details
}
\note{
No notes
}
\examples{
evalHand(mapCards(c("As", "Ks", "Qs", "Js", "Ts", "3c", "5h")))
}
\references{
\url{https://gist.github.com/EluctariLLC/832122}
}

