% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalHand.R
\name{evalMultiHand}
\alias{evalMultiHand}
\title{Evaluate multiple poker hands}
\usage{
evalMultiHand(x)
}
\arguments{
\item{x}{an integer matrix, where each column is a hand, or an integer vector.}
}
\value{
an integer, where higher = better
}
\description{
Returns an integer vector, where higher = a better hand
}
\details{
No details
}
\note{
No notes
}
\examples{
cards <- sapply(1:1e7, function(x) sample(1:52, 7))
system.time(res <- evalMultiHand(cards))
}
\references{
\url{https://gist.github.com/EluctariLLC/832122}
}

