% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cards.R
\name{format.card}
\alias{format.card}
\title{Construct the string representation of a card}
\usage{
\method{format}{card}(x, ...)
}
\arguments{
\item{x}{A card object}

\item{...}{Ignored}
}
\value{
The string representation of the card
}
\description{
Construct the string representation of a card
}
