% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hands.R
\name{format.category}
\alias{format.category}
\title{Construct the string representation of a category}
\usage{
\method{format}{category}(x, ...)
}
\arguments{
\item{x}{A category object}

\item{...}{Ignored}
}
\value{
The string representation of the category
}
\description{
Construct the string representation of a category
}
