% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cards.R
\name{format.hand}
\alias{format.hand}
\title{Re-construct the JSON representation of a hand}
\usage{
\method{format}{hand}(x, ...)
}
\arguments{
\item{x}{A hand object}

\item{...}{Ignored}
}
\value{
A JSON representation of the hand
}
\description{
Re-construct the JSON representation of a hand
}
