% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hands.R
\name{format_cards}
\alias{format_cards}
\title{Format the (subset of) cards used to make the best hand.}
\usage{
format_cards(category)
}
\arguments{
\item{category}{A category object}
}
\value{
A string showing the cards used in the hand
}
\description{
Format the (subset of) cards used to make the best hand.
}
