% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hand.R
\name{generate_hand}
\alias{generate_hand}
\title{Generate a single Texas Hold'Em hand}
\usage{
generate_hand(players, known_p, known_c)
}
\arguments{
\item{players}{Integer scalar, number of players. Must be less than 10.}

\item{known_p}{List of lists of length 2, containing card IDs representing known player hands.}

\item{known_c}{List of length 0 to 5, containing card IDs representing known community cards.}
}
\description{
Randomly generate a Texas Hold'Em poker hand, with optional arguments to tailor
number of players, known player hand(s), and known community cards.
}
