% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hands.R
\name{generate_hands}
\alias{generate_hands}
\title{Generate a single Texas Hold'Em hand}
\usage{
generate_hands(
  n = 1L,
  players = 4L,
  known_p = list(),
  known_c = list(),
  seed = NULL
)
}
\arguments{
\item{n}{Integer scalar, number of hands. Defaults to 1.}

\item{players}{Integer scalar, number of players. Must be less than 10. Defaults to 4.}

\item{known_p}{List of lists of length 2, containing card IDs representing known player hands.
Defaults to empty i.e. all hands unknown.}

\item{known_c}{List of length 0 to 5, containing card IDs representing known community cards.
Default to empty i.e. all community cards unknown.}

\item{seed}{Integer scalar, seed for random number generator. Passed to \code{set.seed}. Defaults
to \code{NULL}.}
}
\description{
Randomly generate a Texas Hold'Em poker hand, with optional arguments to tailor
number of hands, number of players, known player hand(s), known community cards, and the random
number generator seed.
}
