% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{hand}
\alias{hand}
\title{Function: 	hand}
\usage{
hand(players, board)
}
\description{
Author: Ben Greenspan		
Created: May/June 2010
Description:	determine the 7 card hands
Inputs:	@param players :	the hole cards			#matrix[nPlayers, 4]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: suit of card 1 in {1, 2, 3, 4}
											#col3: rank of card 2
											#col4: suit of card 2
		@param board :	the board cards			#vector[5] in {1, 2, ..., 52}
Locals:	nPlayers : number of hands			#integer in {1, 2, ... , 8}
			i : for name in vector
											#numeric in vector
			j : index for hole cards
											#numeric in vector
Outputs:	cards : the 7 card hand
											#matrix[nPlayers, 14]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: suit of card 1 in {1, 2, 3, 4}
											#col3: rank of card 2
											#col4: suit of card 2
											#.
											#.
											#.
											#col13: rank of card 7
											#col14: suit of card 7
}
