% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hands.R
\name{make_rank_vector}
\alias{make_rank_vector}
\title{Make a vector of (integer) category, high card ranks, kicker ranks
that can be used to order hands}
\usage{
make_rank_vector(category)
}
\arguments{
\item{category}{A category object}
}
\value{
An integer vector
}
\description{
Make a vector of (integer) category, high card ranks, kicker ranks
that can be used to order hands
}
