% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalHand.R
\name{mapCards}
\alias{mapCards}
\title{Maps a vector of poker cards to an integer}
\usage{
mapCards(x)
}
\arguments{
\item{x}{a poker card}
}
\description{
2 is low and A is high.  Suits go clubs, diamonds, hearts,
spaces.  E.g. '2c'=1 and 'As'=52
}
\details{
No details
}
\note{
No notes
}
\examples{
mapCards(c("As", "Ks", "Qs", "Js", "Ts", "3c", "5h"))
}

