% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankHands.R
\name{order_hands}
\alias{order_hands}
\title{Ordering permutation of a list of hands.
This is \code{order} for a list of hands.}
\usage{
order_hands(hands, decreasing = TRUE)
}
\arguments{
\item{hands}{A list of hand objects}

\item{decreasing}{Sort in decreasing order (default) or not?}
}
\value{
An integer vector of indices such that \code{hands[order_hands(hands)]}
is in order by rank.
}
\description{
Ordering permutation of a list of hands.
This is \code{order} for a list of hands.
}
