% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cards.R
\name{parse_card}
\alias{parse_card}
\title{Parse a single card}
\usage{
parse_card(card)
}
\arguments{
\item{card}{The text name of a card, e.g. 'JH'}
}
\value{
A two element list containing the numeric card rank
and suit
}
\description{
Parse a single card
}
