% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cards.R
\name{parse_cards}
\alias{parse_cards}
\title{Parse a hand provided as a JSON list of cards}
\usage{
parse_cards(cards)
}
\arguments{
\item{cards}{The JSON input}
}
\value{
A list of cards in decreasing order of rank
}
\description{
Parse a hand provided as a JSON list of cards
}
