% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeR-package.R
\docType{package}
\name{pokeR}
\alias{pokeR}
\alias{pokeR-package}
\title{pokeR: Evaluate poker hands.}
\description{
The pokeR package providers an interface to the 2+2 poker hand evaluator, as
well as some ancillary functions
}

