% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.plumb_package.R
\name{pr.plumb_package}
\alias{pr.plumb_package}
\title{Wrapper function to plumb package and expose API endpoints}
\usage{
pr.plumb_package(...)
}
\arguments{
\item{...}{Arguments to pass to \code{plumber::pr_run}.}
}
\description{
Wrapper function to plumb package and expose API endpoints
}
