% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hands.R
\name{print.category}
\alias{print.category}
\title{Print a category}
\usage{
\method{print}{category}(x, ...)
}
\arguments{
\item{x}{A category object}

\item{...}{Passed to \code{print.default}}
}
\description{
Print a category
}
