% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankHands.R
\name{rank_hands}
\alias{rank_hands}
\title{Compute ranks (ordinal order) for a sorted list of hands.
This is \code{rank(ties.method='min', ...)} for an already sorted list}
\usage{
rank_hands(hands)
}
\arguments{
\item{hands}{A list of hand objects}
}
\value{
An integer vector of ranks
}
\description{
Compute ranks (ordinal order) for a sorted list of hands.
This is \code{rank(ties.method='min', ...)} for an already sorted list
}
