% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunExample.R
\name{runExample}
\alias{runExample}
\title{Run a simple Shiny app demonstrating the poker hand parsing and ranking.}
\usage{
runExample()
}
\description{
Run a simple Shiny app demonstrating the poker hand parsing and ranking.
}
