% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{scorer}
\alias{scorer}
\title{Function: 	scorer}
\usage{
scorer(cardsRow)
}
\description{
Author: Ben Greenspan		
Created: December 15, 2010
Updated: December 17, 2010
Description:	returns the ranking of one hand
Inputs:		@param cardsRow : one 7 card hand
											#vector[14]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: suit of card 1 in {1, 2, 3, 4}
											#col3: rank of card 2
											#col4: suit of card 2
											#.
											#.
											#.
											#col13: rank of card 7
											#col14: suit of card 7
Locals:		ranks : the indices for rank	#vector
			suits : the indices for suit	#vector
				sortedSuits : the sorted suits
											#vector
				sortedRanks : the sorted suits
											#vector
			suitValues : the unique values of equal suits
											#vector
											#col1: lowest unique suit
											#.
											#.
											#.
											#coln: highest unique suit
			suitLengths : the lengths of equal suits
											#vector
											#col1: length of lowest rank
											#.
											#.
											#.
											#coln: length of highest rank
			rankValues : the unique values of equal ranks
											#vector
											#col1: lowest unique rank
											#.
											#.
											#.
											#coln: highest unique rank
			rankLengths : the lengths of equal ranks
											#vector
											#col1: length of lowest rank
											#.
											#.
											#.												
			coln: length of highest rank
			j : for name in vector
											#numeric in vector
			k : the length of the length of unique ranks
											#integer
			kgt4 : the positive difference between k and 4
											#integer
			straight : create an indicator variable for the presence of a straight
											#boolean
			yTemp : the 7 card hand transformed to number
											#vector
											#col1: the number of card 1 in {1, 2, ... , 52}
											#.
											#.
											#.
											#col7: the number of card 7
			straightFlushCounter : a counter for the number of sequential cards of the same rank
											#integer
Outputs:		@return ranking : the ranks of the hand
											#integer in {1, 2, ... , 9}
}
