% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{showdown}
\alias{showdown}
\title{Function: 	showdown}
\usage{
showdown(cards)
}
\description{
Author: Ben Greenspan		
Created: December 15, 2010
Description:	returns the ranking of the hands
Inputs: 		@param cards :	the 7 card hand
											#matrix[nPlayers, 14]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: suit of card 1 in {1, 2, 3, 4}
											#col3: rank of card 2
											#col4: suit of card 2
Locals:				
Outputs: 	#score : the score of the hand in absolute terms
											#vector[nPlayers]
											# 9 = Straight Flush
											# 8 = Four of a Kind
											# 7 = Full House
											# 6 = Flush
											# 5 = Straight
											# 4 = Three of a Kind
											# 3 = Two Pair
											# 2 = One Pair
											# 1 = High Card
}
