% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{tiebreaker}
\alias{tiebreaker}
\title{Function: 	tiebreaker}
\usage{
tiebreaker(nPlayers, cards, score)
}
\description{
Description:	in case of >= 2 similar scores, determines a winner
Inputs: 		@param nPlayers : number of hands			#integer in {1, 2, ... , 8}
			@param score : the score of the hand in absolute terms
											#vector[nPlayers]
											# 9 = Straight Flush
											# 8 = Four of a Kind
											# 7 = Full House
											# 6 = Flush
											# 5 = Straight
											# 4 = Three of a Kind
											# 3 = Two Pair
											# 2 = One Pair
											# 1 = High Card
			@param cards :	the 7 card hand
											#matrix[nPlayers, 14]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: suit of card 1 in {1, 2, 3, 4}
											#col3: rank of card 2
											#col4: suit of card 2
Locals:	
Outputs: 	#winner : the absolute position of the winner[s]
											#vector[]
}
