% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{transformToNumber}
\alias{transformToNumber}
\title{Function: 	transformToNumber}
\usage{
transformToNumber(rank, suit)
}
\description{
Author: Ben Greenspan		
Created: May/June 2010
Description:	transforms a rank and suit into a number
Inputs:		@param rank: rank of card y 
											#integer in {2, ... , 14}
											# 2 = deuce
											# .
											# .
											# .
											# 11 = jack
											# 12 = queen
											# 13 = king
											# 14 = ace 
			@param suit: suit of card y 
											#integer in {1, 2, 3, 4}
											# 1 = ♠
											# 2 = ♣
											# 3 = ♥
											# 4 = ♦
Outputs:		@return y : number corresponding to card
											#integer in {1, 2, ... , 52}
}
