% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{transformToRank}
\alias{transformToRank}
\title{Function: 	transformToRank}
\usage{
transformToRank(y)
}
\description{
Author: Ben Greenspan		
Created: May/June 2010
Description:	transforms a number into a rank
Inputs:		@param y : number corresponding to card
											#integer in {1, 2, ... , 52}
Outputs:		@return rank: rank of card y 
											#integer in {2, ... , 14}
											# 2 = deuce
											# .
											# .
											# .
											# 11 = jack
											# 12 = queen
											# 13 = king
											# 14 = ace
}
