% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{transformToRelative}
\alias{transformToRelative}
\title{Function: 	transformToRelative}
\usage{
transformToRelative(nPlayers, position, j)
}
\description{
Author: Ben Greenspan		
Created: May/June 2010
Description:	transforms an absolute (seat at the table) position into a relative (seats behind the dealer) position
Inputs:		@param nPlayers : number of hands to deal 
											#integer in {1, 2, ... , 8}
			@param position : dealer position 		#integer in {1, 2, ..., nPlayers}
			@param j : absolute position of a player
											#integer in {1, 2, ... , nPlayers}
Outputs:		@return k : relative position of a player
											#integer in {1, 2, ... , nPlayers}
}
