% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{transformToSuit}
\alias{transformToSuit}
\title{transformToSuit}
\usage{
transformToSuit(y)
}
\arguments{
\item{y}{number corresponding to card as
integer in \{1, 2, ... , 52\} \cr}
}
\value{
\tabular{ll}{suit: suit of card y as
											integer in \{1, 2, 3, 4\} \tab \cr \tab
											 1 = spade \cr \tab
											 2 = club \cr \tab
											 3 = heart \cr \tab
											 4 = diamond}
}
\description{
Determines the suit of a card. \cr
}
\examples{
transformToSuit(1)
transformToSuit(13)
transformToSuit(14)
transformToSuit(26)
}
