% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what_hand.R
\name{what_hand}
\alias{what_hand}
\title{Assesses the strength of a 5-card poker hand.}
\usage{
what_hand(hand)
}
\arguments{
\item{hand}{5-card poker hand to evaluate.}
}
\value{
What hand the cards make.
}
\description{
Assesses the strength of a 5-card poker hand.
}
\examples{
what_hand(deal_hand())
}
