% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{assignToBoard}
\alias{assignToBoard}
\title{assignToBoard}
\usage{
assignToBoard(y)
}
\arguments{
\item{y}{cards dealt as                    vector[2*nPlayers+3] in \{1, 2, ..., 52\}}
}
\value{
board :	the board cards	as		vector[5] in \{1, 2, ..., 52\}
}
\description{
Deal 3 community cards.
}
\examples{
assignToBoard(1:23)
assignToBoard(c(1:17,24,48:52))
}
