% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hands.R
\name{categorize_hand}
\alias{categorize_hand}
\title{Find the category of a hand.
Will work with hands of any size up to 5}
\usage{
categorize_hand(hand)
}
\arguments{
\item{hand}{A hand}
}
\value{
A category object
}
\description{
Find the category of a hand.
Will work with hands of any size up to 5
}
