% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{CGIplayers}
\alias{CGIplayers}
\title{Function: 	CGIplayers}
\usage{
CGIplayers(time, alias, position, cards, score)
}
\description{
Author: Ben Greenspan		
Created: May/June 2010
Description:	computer graphics
Inputs: 		@param time : the current round			#integer in {1, 2, 3, 4}
											# 1 = pre-flop
											# 2 = flop
											# 3 = turn
											# 4 = river
			@param alias : names of players			#vector[nPlayers]
			@param position : dealer position 		#integer in {1, 2, ..., nPlayers}
			@param cards : the 7 card hand
											#matrix[nPlayers, 14]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: suit of card 1 in {1, 2, 3, 4}
											#col3: rank of card 2
											#col4: suit of card 2
											#.
											#.
											#.
											#col13: rank of card 7
											#col14: suit of card 7
			@param score : the score of the hand in absolute terms
											#vector[nPlayers]
											# 9 = Straight Flush
											# 8 = Four of a Kind
											# 7 = Full House
											# 6 = Flush
											# 5 = Straight
											# 4 = Three of a Kind
											# 3 = Two Pair
											# 2 = One Pair
											# 1 = High Card
Locals:	#nPlayers : number of hands			#integer in {1, 2, ... , 8}
		#width : width of the plot			#integer in {20, 30, ... , 10*nPlayers}
		#height : height of the plot		#integer in {10}
		#X1 : x-coordinate for Players		#real number in [0, width]
		#Y1 : y-coordinate for Players		#real number in [0, height]
		#X2 : x-coordinate for Cards		#real number in [0, width]
		#Y2 : y-coordinate for Cards		#real number in [0, height]
		#X3 : x-coordinate for Community	#real number in [0, width]
		#Y3 : y-coordinate for Community	#real number in [0, height]
		#X4 : x-coordinate for outline		#real number in [0, width]
		#Y4 : y-coordinate for outline		#real number in [0, height]
Outputs:
}
