% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotFlushRanker}
\alias{dotFlushRanker}
\title{dotFlushRanker}
\usage{
dotFlushRanker(cardsRow)
}
\arguments{
\item{cardsRow}{: \tabular{ll}{one 7 card hand as
vector[14] \tab \cr \tab
col1: rank of card 1 in \{2, ... , 14\} \cr \tab
col2: suit of card 1 in \{1, 2, 3, 4\} \cr \tab
col3: rank of card 2 \cr \tab
col4: suit of card 2 \cr \tab
. \cr \tab
. \cr \tab
. \cr \tab
col13: rank of card 7 \cr \tab
col14: suit of card 7}}
}
\value{
\tabular{ll}{flushRank : the rank of 5 high cards in flush as
											vector[5] \tab \cr \tab
											col1: suit of card 1 in \{2, ... , 14\} \cr \tab
											. \cr \tab
											. \cr \tab
											. \cr \tab
											col5: suit of card n in \{2, ... , 14\}}
}
\description{
Return the ranks of the 5 highest cards in the flush.
}
\examples{
dotFlushRanker(c(2,1,2,2,5,2,7,2,8,2,9,2,11,1))
dotFlushRanker(c(2,1,2,2,5,2,7,2,8,2,9,2,11,2))
}
