% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotFourOfAKind}
\alias{dotFourOfAKind}
\title{dotFourOfAKind}
\usage{
dotFourOfAKind(nPlayers, cards, score)
}
\arguments{
\item{nPlayers}{number of hands as            integer in \{2, ... , 9\}}

\item{cards}{\tabular{ll}{the 7 card hand as
matrix[nPlayers, 14] \tab \cr \tab
col1: rank of card 1 in \{2, ... , 14\} \cr \tab
col2: suit of card 1 in \{1, 2, 3, 4\} \cr \tab
col3: rank of card 2 \cr \tab
col4: suit of card 2 \cr \tab
. \cr \tab
. \cr \tab
. \cr \tab
col13: rank of card 7 \cr \tab
col14: suit of card 7}}

\item{score}{\tabular{ll}{the score of the hand in absolute terms as
vector[nPlayers] \tab \cr \tab
 9 = Straight Flush \cr \tab
 8 = Four of a Kind \cr \tab
 7 = Full House \cr \tab
 6 = Flush \cr \tab
 5 = Straight \cr \tab
 4 = Three of a Kind \cr \tab
 3 = Two Pair \cr \tab
 2 = One Pair \cr \tab
 1 = High Card}}
}
\value{
winner : absolute position of the winner as
											vector
}
\description{
Determines the player with the highest hand (i.e., four of a kind and kicker) with score of 8.
}
\details{
Locals: \cr
ranks : the indices for rank as	vector \cr
\tabular{ll}{temp : a copy of cards with ranks only as
											matrix[nPlayers, 5] \tab \cr \tab
											col1: rank of card 1 in \{2, ... , 14\} \cr \tab
											col2: rank of card 2 \cr \tab
											. \cr \tab
											. \cr \tab
											. \cr \tab
											col7: rank of card 7} \cr
\tabular{ll}{fourOfAKindRanks : a matrix for the rank of the quads and the kicker as
											matrix[nPlayers,2] \tab \cr \tab
											col1: rank of high set \cr \tab
											col2: rank of high pair} \cr
}
\examples{
dotFourOfAKind(3,matrix(c(14, 10, 5, 1, 2, 1, 14, 9, 7, 2, 2, 2, 4, 4, 4, 3, 3, 3, 8, 8, 8, 3, 3, 3, 13, 13, 13, 3, 3, 3, 14, 14, 14, 3, 3, 3, 14, 14, 14, 4, 4, 4), byrow=FALSE,nrow=3,ncol=14),c(8,2,2))
dotFourOfAKind(4,matrix(c(3, 4, 5, 1, 1, 1, 8, 9, 10, 1, 1, 1, 14, 14, 14, 1, 1, 1, 14, 14, 14, 2, 2, 2, 7, 7, 7, 3, 3, 3, 14, 14, 14, 3, 3, 3, 14, 14, 14, 4, 4, 4), byrow=FALSE,nrow=3,ncol=14),c(8,8,8))
dotFourOfAKind(4,matrix(c(3, 4, 5, 1, 1, 1, 8, 9, 10, 1, 1, 1, 14, 14, 14, 1, 1, 1, 14, 14, 14, 2, 2, 2, 11, 11, 11, 3, 3, 3, 14, 14, 14, 3, 3, 3, 14, 14, 14, 4, 4, 4), byrow=FALSE,nrow=3,ncol=14),c(8,8,8))
}
