% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotFourOfAKindRanker}
\alias{dotFourOfAKindRanker}
\title{Function: 	dotFourOfAKindRanker}
\usage{
dotFourOfAKindRanker(oneHand)
}
\description{
Author: Ben Greenspan		
Created: December 15, 2010
Updated: December 17, 2010
Description:	returns the rank of the four of a kind and the kicker
Inputs:		@param oneHand : the ranks of one 7 card hand
											#vector[7]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: rank of card 2
											#.
											#.
											#.
											#col7: rank of card 7
Locals:		values : the unique values of equal ranks
											#vector
											#col1: highest unique rank
											#.
											#.
											#.
											#coln: lowest unique rank
			lengths : the lengths of equal ranks
											#vector
											#col1: length of highest rank
											#.
											#.
											#.
											#coln: length of lowest rank
Outputs:		@return fourOfAKindRank : the ranks of the quads and the high kicker
											#vector
											#col1: the rank of the quads
											#col2: the rank of the kicker
}
