% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotFullHouseRanker}
\alias{dotFullHouseRanker}
\title{dotFullHouseRanker}
\usage{
dotFullHouseRanker(oneHand)
}
\arguments{
\item{oneHand}{: \tabular{ll}{the ranks of one 7 card hand as
vector[7] \tab \cr \tab
col1: rank of card 1 in \{2, ... , 14\} \cr \tab
col2: rank of card 2 \cr \tab
. \cr \tab
. \cr \tab
. \cr \tab
col7: rank of card 7}}
}
\value{
\tabular{ll}{fullHouseRank : the ranks of the high set and the high pair as
											vector \tab \cr \tab
											col1: the rank of the top set \cr \tab
											col2: the rank of the top pair}
}
\description{
Determine the rank of the top set and the top pair.
}
\examples{
dotFullHouseRanker(c(2,2,2,5,5,8,9))
dotFullHouseRanker(c(2,2,5,5,5,8,9))
dotFullHouseRanker(c(2,2,5,5,5,8,8))
}
