% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotHighcard}
\alias{dotHighcard}
\title{dotHighcard}
\usage{
dotHighcard(cards)
}
\arguments{
\item{cards}{: \tabular{ll}{ the 7 card hand as
matrix[nPlayers, 14] \tab \cr \tab
col1: rank of card 1 in \{2, ... , 14\} \cr \tab
col2: suit of card 1 in \{1, 2, 3, 4\} \cr \tab
col3: rank of card 2 \cr \tab
col4: suit of card 2 \cr \tab
. \cr \tab
. \cr \tab
. \cr \tab
col13: rank of card 7 \cr \tab
col14: suit of card 7}}
}
\value{
winner : absolute position of the winner as
											vector
}
\description{
Determine the player(s) with a high card hand.
}
\examples{
dotHighcard(matrix(c(2,1,14,2,5,3,6,4,7,1,13,2,14,3,2,3,3,4,5,1,6,2,7,3,13,4,14,1),2,14,byrow=TRUE))
dotHighcard(matrix(c(2,1,3,2,5,3,6,4,7,1,13,2,14,3,2,3,3,4,5,1,6,2,7,3,13,4,14,1),2,14,byrow=TRUE))
}
\seealso{
\code{\link{dotHighcardCompare}}
}
