% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotHighcardCompare}
\alias{dotHighcardCompare}
\title{dotHighcardCompare}
\usage{
dotHighcardCompare(rankMatrix)
}
\arguments{
\item{rankMatrix}{: \tabular{ll}{the ranks from the 7 card hand as
matrix[nPlayers, 7] \tab \cr \tab
col1: rank of card 1 in \{2, ... , 14\} \cr \tab
col2: rank of card 2 \cr \tab
. \cr \tab
. \cr \tab
. \cr \tab
col7: rank of card 7}}
}
\value{
winner : absolute position of the winner as
											vector
}
\description{
Determine the player(s) with the high card.
}
\examples{
dotHighcardCompare(matrix(c(2,4,5,6,7,13,14,2,3,5,6,7,13,14),2,7,byrow=TRUE))
dotHighcardCompare(matrix(c(2,3,5,6,7,13,14,2,3,5,6,7,13,14),2,7,byrow=TRUE))
}
