% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotPair}
\alias{dotPair}
\title{Function: 	dotPair}
\usage{
dotPair(nPlayers, cards, score)
}
\description{
Author: Ben Greenspan		
Created: December 14, 2010
Updated: December 17, 2010
Description:	determines the players with the highest pair and kicker cards.
Inputs:		@param nPlayers : number of hands		#integer in {1, 2, ... , 8}
			@param cards : the 7 card hand
											#matrix[nPlayers, 14]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: suit of card 1 in {1, 2, 3, 4}
											#col3: rank of card 2
											#col4: suit of card 2
											#.
											#.
											#.
											#col13: rank of card 7
											#col14: suit of card 7
			@param score : the score of the hand in absolute terms
											#vector[nPlayers]
											# 9 = Straight Flush
											# 8 = Four of a Kind
											# 7 = Full House
											# 6 = Flush
											# 5 = Straight
											# 4 = Three of a Kind
											# 3 = Two Pair
											# 2 = One Pair
											# 1 = High Card
Locals:		ranks : the indices for rank	#vector
			temp : a copy of cards with ranks only
											#matrix[nPlayers, 7]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: rank of card 2
											#.
											#.
											#.
											#col7: rank of card 7
			pairRanks : the rank of each pair
											#vector[nPlayers]
			kickers : the kicker cards with respect to a pair, i.e. 3 highest nonpaired cards
											#matrix[nPlayers,3]
			step : the step for comparison of highest card
											#integer
			Tie : the positions of the players with Tied kicker cards
											#vector
Outputs:		@return winner : absolute position of the winner
											#vector
}
