% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotPairRanker}
\alias{dotPairRanker}
\title{dotPairRanker}
\usage{
dotPairRanker(oneHand)
}
\arguments{
\item{oneHand}{: \tabular{ll}{a sorted hand with ranks only as
vector[7] \tab \cr \tab
col1: rank of card 1 in \{2, ... , 14\} \cr \tab
col2: rank of card 2 \cr \tab
. \cr \tab
. \cr \tab
. \cr \tab
col7: rank of card 7}}
}
\value{
pairRank : the rank of the pair as
											vector
}
\description{
Determine the rank of the pair.  Notes: dotPairRanker requires a hand with a score of 2 (i.e., a pair). This functions works best when ranks are sorted in decreasing order.
}
\examples{
dotPairRanker(c(2,2,5,6,7,13,14))
}
