% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotStraight}
\alias{dotStraight}
\title{Function: 	dotStraight}
\usage{
dotStraight(cards, score)
}
\description{
Author: Ben Greenspan		
Created: December 14, 2010
Description:	determines the player with the highest straight
Inputs:		@param cards : the 7 card hand
											#matrix[nPlayers, 14]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: suit of card 1 in {1, 2, 3, 4}
											#col3: rank of card 2
											#col4: suit of card 2
											#.
											#.
											#.
											#col13: rank of card 7
											#col14: suit of card 7
			@param score : the score of the hand in absolute terms
											#vector[nPlayers]
											# 9 = Straight Flush
											# 8 = Four of a Kind
											# 7 = Full House
											# 6 = Flush
											# 5 = Straight
											# 4 = Three of a Kind
											# 3 = Two Pair
											# 2 = One Pair
											# 1 = High Card
Locals:		ranks : the indices for rank	#vector
			temp : a copy of cards with ranks only
											#matrix[nPlayers, 7]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: rank of card 2
											#.
											#.
											#.
											#col7: rank of card 7
			maxStraight : the high card in each straight
											#vector[nPlayers]
Outputs:		@return winner : absolute position of the winner
											#vector
}
