% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotTestDealer}
\alias{dotTestDealer}
\title{dotTestDealer}
\usage{
dotTestDealer(nPlayers, position, holeCards)
}
\arguments{
\item{nPlayers}{number of hands to deal as 
integer in \{2, ... , 9\}}

\item{position}{dealer position     as integer in \{2, ..., nPlayers\}}

\item{holeCards}{the hand of player 1 as vector[2] in \{1, 2, ..., 52\}}
}
\value{
y : cards dealt in hole as		vector[nCards] in \{1, 2, ..., 52\}
}
\description{
Assume player 1 already has cards.  For remaining players, generate Player+Community cards = 2x(nPlayers-1)+5 cards. \cr
}
\examples{
dotTestDealer(9,9,c(1,52))
dotTestDealer(9,5,c(1,52))
dotTestDealer(5,2,c(3,42))
}
