% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotTripRanker}
\alias{dotTripRanker}
\title{Function: 	dotTripRanker}
\usage{
dotTripRanker(oneHand)
}
\description{
Author: Ben Greenspan		
Created: December 17, 2010
Description:	This function requires a hand with a score of 2. The function determines the rank of the pair.  
Inputs:		@param oneHand : a sorted hand with ranks only.
											#vector[7]
											#col1: rank of card 1 in {2, ... , 14}
											#col2: rank of card 2
											#.
											#.
											#.
											#col7: rank of card 7
			@param rankValues : the unique values of equal ranks
											#vector
											#col1: lowest unique rank
											#.
											#.
											#.
											#coln: highest unique rank
			@param rankLengths : the lengths of equal ranks
											#vector
											#col1: length of lowest rank
											#.
											#.
Outputs:		@return pairRank : the rank of the pair
											#vector
}
