% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{dotTwoPairRanker}
\alias{dotTwoPairRanker}
\title{dotTwoPairRanker}
\usage{
dotTwoPairRanker(oneHand)
}
\arguments{
\item{oneHand}{\tabular{ll}{a sorted hand with ranks only as
vector[7] \tab \cr \tab
col1: rank of card 1 in \{2, ... , 14\} \cr \tab
col2: rank of card 2 \cr \tab
. \cr \tab
. \cr \tab
. \cr \tab
col7: rank of card 7}}
}
\value{
pairRank : the rank of the pair as
											vector
}
\description{
Determines the ranks of the two pairs.  Notes: dotTwoPairRanker requires a hand with a score of 3 (i.e., two pairs). This functions works best when ranks are sorted in decreasing order. \cr
}
\details{
Locals: \cr
\tabular{ll}{rankValues: the unique values of equal ranks as
											vector \tab \cr \tab
											col1: lowest unique rank \cr \tab
											. \cr \tab
											. \cr \tab
											. \cr \tab
											coln: highest unique rank} \cr
\tabular{ll}{rankLengths: the lengths of equal ranks as
											vector \tab \cr \tab
											col1: length of lowest rank \cr \tab
											. \cr \tab
											.} \cr
}
\examples{
dotTwoPairRanker(c(9,7,5,3,3,2,2))
dotTwoPairRanker(c(9,5,5,3,3,2,2))
}
