% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalHand.R
\name{evalMultiHand}
\alias{evalMultiHand}
\title{Evaluate multiple poker hands}
\usage{
evalMultiHand(x)
}
\arguments{
\item{x}{An integer vector.}
}
\value{
an integer, where higher = better
}
\description{
Returns an integer vector, where higher = a better hand
}
\details{
No details
}
\note{
No notes
}
\examples{
n <- 2e7 #20 million hands
cards <- sample(1:52, 7*n, replace=TRUE)
t1 <- system.time(res <- evalMultiHand(cards))
hps <- prettyNum(n / t1[['elapsed']],big.mark=",",scientific=FALSE)
print(paste("evaluated", hps, "poker hands per second"))
}
\references{
\url{https://gist.github.com/EluctariLLC/832122}
}

