% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hands.R
\name{format_cards}
\alias{format_cards}
\title{Format the cards used to show the hand.}
\usage{
format_cards(category)
}
\arguments{
\item{category}{A category object}
}
\value{
A string showing the cards used in the hand
}
\description{
Format the cards used to show the hand.
}
