% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{testRoundOfPoker}
\alias{testRoundOfPoker}
\title{testRoundOfPoker}
\usage{
testRoundOfPoker()
}
\value{
Outputs a plot window showing the cards dealt as well as prints to the console the summary text, i.e., each hand's score and the winner.
}
\description{
Run a test round of poker.
}
\examples{
testRoundOfPoker()
}
\seealso{
\code{\link{deal}}, \code{\link{assignToPlayers}}, \code{\link{assignToBoard}}, \code{\link{hand}}, \code{\link{showdown}}, \code{\link{tiebreaker}}, and \code{\link{cgiPlayers}}
}
