% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{transformToAbsolute}
\alias{transformToAbsolute}
\title{Function: 	transformToAbsolute}
\usage{
transformToAbsolute(nPlayers, position, k)
}
\description{
Author: Ben Greenspan		
Created: May/June 2010
Description:	transforms a relative position (seats behind the dealer) into an absolute position (seat at the table) 
Inputs:		@param nPlayers : number of hands to deal 
											#integer in {1, 2, ... , 8}
			@param position : dealer position 		#integer in {1, 2, ..., nPlayers}
			@param k : relative position of a player
											#integer in {1, 2, ... , nPlayers} 
Outputs:		@return j : absolute position of a player
											#integer in {1, 2, ... , nPlayers}
}
