% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{transformToRelative}
\alias{transformToRelative}
\title{transformToRelative}
\usage{
transformToRelative(nPlayers, position, j)
}
\arguments{
\item{nPlayers}{number of hands to deal as
integer in \{2, ... , 9\}}

\item{position}{dealer position as        integer in \{2, ..., nPlayers\}}

\item{j}{absolute position of a player as
integer in \{1, 2, ... , nPlayers\}}
}
\value{
k : relative position of a player as
											integer in \{1, 2, ... , nPlayers\}
}
\description{
Transforms an absolute position (i.e., seat at the table) into a relative position (i.e., seats behind the dealer)
}
\examples{
transformToRelative(9,9,9)
transformToRelative(9,9,8)
transformToRelative(9,1,9)
transformToRelative(9,5,2)
}
