% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{transformToSuit}
\alias{transformToSuit}
\title{Function: 	transformToSuit}
\usage{
transformToSuit(y)
}
\description{
Author: Ben Greenspan		
Created: May/June 2010
Description:	transforms a number into a suit
Inputs:		@param y : number corresponding to card
											#integer in {1, 2, ... , 52}
Outputs:		@return suit: suit of card y 
											#integer in {1, 2, 3, 4}
											# 1 = ♠
											# 2 = ♣
											# 3 = ♥
											# 4 = ♦
}
